umbra/requirements.txt
Flask==3.1.0
Flask-APScheduler==1.13.1
pyserial==3.5